requirements.txt
Flask
Flask-Script
Flask-SQLAlchemy
wtforms
flask-mongoengine
nodeenv
mysql-python